Refactor downloads channel and provider response.
The channel design is now consists of worker and jobs. Downloader
is a worker (which implemented with simply http.Get) and a job
is a provider.ProviderItem type where worker downloads ProviderItem.Link
and save it into *out + ProviderItem.Filename. Results channel is
a string representing the result of downloading.